MuseDjango/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.9b0
boto3==1.18.48
botocore==1.21.48
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.8.0
django-crontab==0.7.1
django-storages==1.11.1
django-taggit==1.5.1
django-taggit-serializer==0.1.7
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docker==5.0.2
idna==3.2
jmespath==0.10.0
kombu==5.1.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
prompt-toolkit==3.0.21
pycparser==2.20
PyJWT==2.2.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
